refactor(store): move mapped file owner to local
