+++ rocketmq-store/Cargo.toml
@@ -96,7 +96,6 @@ futures-util = "0.3.32"
 rand         = { workspace = true }
 uuid         = { workspace = true }
 anyhow       = { workspace = true }
-page_size    = "0.6.0"
 fs2          = "0.4"
 
 dashmap   = { workspace = true }
